+++ packages/jest-types/src/Config.ts
@@ -350,7 +350,7 @@ export type ProjectConfig = {
   displayName?: DisplayName;
   errorOnDeprecated: boolean;
   extensionsToTreatAsEsm: Array<Path>;
-  extraGlobals: Array<keyof NodeJS.Global>;
+  extraGlobals: Array<keyof typeof globalThis>;
   filter?: Path;
   forceCoverageMatch: Array<Glob>;
   globalSetup?: string;