+++ packages/jest-worker/src/workers/NodeThreadsWorker.ts
@@ -61,7 +61,6 @@ export default class ExperimentalWorker implements WorkerInterface {
   initialize(): void {
     this._worker = new Worker(path.resolve(__dirname, './threadChild.js'), {
       eval: false,
-      // @ts-expect-error: added in newer versions
       resourceLimits: this._options.resourceLimits,
       stderr: true,
       stdout: true,
@@ -76,7 +75,7 @@ export default class ExperimentalWorker implements WorkerInterface {
         silent: true,
         ...this._options.forkOptions,
       },
-    });
+    } as import('worker_threads').WorkerOptions);
 
     if (this._worker.stdout) {
       if (!this._stdout) {